// types/type.d.ts
declare interface GoogleInputProps {
  icon?: string;
  initialLocation?: string;
  containerStyle?: string;
  placeholder?: string;
  textInputBackgroundColor?: string;
  handlePress: ({
    latitude,
    longitude,
    address,
  }: {
    latitude: number;
    longitude: number;
    address: string;
  }) => void;
  onTextChange?: (text: string) => void;
  autoFocus?: boolean;
  returnKeyType?: "search" | "done" | "go" | "next" | "send";
  onSubmitEditing?: (event: { nativeEvent: { text: string } }) => void;
}